requirements.txt
flask==2.2.0
flask-sqlalchemy==3.0.0
pdfminer.six==20211012
python-docx==0.8.11
requests==2.26.0
